fix: is sumoling exists condition
