+++ src/lib/appsumo/actions/is-sumo-ling-exists.ts
@@ -1,0 +1,10 @@
+import { getOrganizationsCollection } from "~/lib/server/collections";
+
+export const isSumolingExists = async (uuid: string) => {
+  const { size } = await getOrganizationsCollection()
+    .where("sumolingUUID", "==", uuid)
+    .limit(1)
+    .get();
+
+  return size > 0;
+};
